Refactored oecsm packages and schema-related classes for clarity and consistency
